Update TableComponent.tsx to use `deepClone` function instead of spread syntax
When using the table within a block editor, having created a table, left the document and returned, adding a new row/s would result in an 'can't define property "x": "obj" is not extensible' error in the console.

Both the `addRows` and `addRowAt` functions were using spread syntax when assigning the `newValue` const, rather than using the `deepClone` function which is used by all other functions, which resulted in the error.